prove/requirements.txt
# Core dependencies
Flask>=3.0.0
Werkzeug>=3.0.1
Jinja2>=3.1.3
MarkupSafe>=2.1.3

# Database and forms
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.23
Flask-WTF>=1.2.1
WTForms>=3.1.1
email-validator>=2.1.0.post1

# Security
Flask-Login>=0.6.3
python-dotenv>=1.0.0

# Development tools
pytest>=7.4.3
python-dateutil>=2.8.2